Safari-609.1.11.1/LayoutTests/http/tests/performance/performance-resource-timing-entries-iterable-expected.txt

Test that PerformanceEntryList (window.performance.getEntries()) is iterable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resources is an instance of Array
PASS resources.length is defined.
PASS resources.length is 1
PASS resources[0] is defined.
PASS entryTypes[0] is "resource"
PASS successfullyParsed is true

TEST COMPLETE
